# docs/swagger.yaml
basePath: /
definitions:
  api.adminStatsResponse:
    properties:
      total_categories:
        type: integer
      total_transactions:
        type: integer
      total_users:
        type: integer
    type: object
  api.adminUserResponse:
    properties:
      avatar_url:
        type: string
      created_at:
        $ref: '#/definitions/pgtype.Timestamptz'
      email:
        type: string
      google_id:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
      updated_at:
        $ref: '#/definitions/pgtype.Timestamptz'
    type: object
...
  api.listUsersResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
      users:
        items:
          $ref: '#/definitions/api.adminUserResponse'
        type: array
    type: object
...
  api.updateUserRoleRequest:
    properties:
      role:
        enum:
        - ADMIN
        - USER
        type: string
    required:
    - role
    type: object
...
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    format: int32
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
...
  pgtype.Timestamptz:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        type: string
      valid:
        type: boolean
    type: object
host: localhost:3000
info:
  contact: {}
  description: API untuk aplikasi pencatatan keuangan pribadi
  title: Catat API
  version: "1.0"
paths:
  /admin/stats:
    get:
      consumes:
      - application/json
      description: Mendapatkan statistik admin (total user, transaksi, kategori)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.adminStatsResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get admin statistics (Admin only)
      tags:
      - Admin
  /admin/users:
    get:
      consumes:
      - application/json
      description: Mendapatkan daftar semua user dengan pagination dan pencarian
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: 'Number of items per page (max: 100)'
        in: query
        name: limit
        type: integer
      - description: Search by name or email
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.listUsersResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all users (Admin only)
      tags:
      - Admin
  /admin/users/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus user berdasarkan ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete user (Admin only)
      tags:
      - Admin
    get:
      consumes:
      - application/json
      description: Mendapatkan detail user berdasarkan ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.adminUserResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user by ID (Admin only)
      tags:
      - Admin
  /admin/users/{id}/role:
    put:
      consumes:
      - application/json
      description: Mengubah role user (ADMIN atau USER)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update role request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.updateUserRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.adminUserResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update user role (Admin only)
      tags:
      - Admin
...
  /users/google/callback:
    get:
      consumes:
      - application/json
      description: Endpoint callback untuk Google OAuth, memproses autentikasi dan
        membuat/login user
      parameters:
      - description: Authorization code dari Google
        in: query
        name: code
        required: true
        type: string
      - description: State parameter
        in: query
        name: state
        type: string
      produces:
      - application/json
      responses:
        "307":
          description: Redirect ke frontend dengan error message
          schema:
            type: string
      summary: Callback Google OAuth
      tags:
      - Authentication
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"